Revert "merge logger from Mirror (#224)"
This reverts commit 5b0bc79c0dbfcec69c3dd135a0e0b34b8ebb5488.
