+++ Assets/Mirror/Editor/LogLevelWindow.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f28def2148ed5194abe70af012a4e3e0
+guid: c3dbf48190d77d243b87962a82c3b164
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
