--- Assets/Mirror/Editor/Logging.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4d97731cd74ac8b4b8aad808548ef9cd
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
