--- Assets/Mirror/Editor/Logging/LogLevelWindow.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c3dbf48190d77d243b87962a82c3b164
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
