--- Assets/Mirror/Editor/Logging/LogLevelsGUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9d6ce9d62a2d2ec4d8cef8a0d22b8dd2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
