--- Assets/Mirror/Editor/Logging/LogSettingsEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8f4ecb3d81ce9ff44b91f311ee46d4ea
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
