--- Assets/Mirror/Editor/Logging/NetworkLogSettingsEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 37fb96d5bbf965d47acfc5c8589a1b71
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
