--- Assets/Mirror/Editor/ScriptableObjectUtility.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4d54a29ddd5b52b4eaa07ed39c0e3e83
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
