--- Assets/Mirror/LogSettings.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e200f82c971c8354481d2782f7182939
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
