--- Assets/Mirror/Runtime/Logging.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 28c6aa26ff0ab0441a7c937e6c9ecd09
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
