--- Assets/Mirror/Runtime/Logging/ConsoleColorLogHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2a9618569c20a504aa86feb5913c70e9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
