--- Assets/Mirror/Runtime/Logging/EditorLogSettingsLoader.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a39aa1e48aa54eb4e964f0191c1dcdce
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
