--- Assets/Mirror/Runtime/Logging/LogSettings.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 633889a39717fde4fa28dd6b948dfac7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
