--- Assets/Mirror/Runtime/Logging/NetworkHeadlessLogger.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c7627623f2b9fad4484082517cd73e67
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {fileID: 2800000, guid: 7453abfe9e8b2c04a8a47eb536fe21eb, type: 3}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
