--- Assets/Mirror/Runtime/Logging/NetworkLogSettings.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ac6e8eccf4b6f4dc7b24c276ef47fde8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {fileID: 2800000, guid: 7453abfe9e8b2c04a8a47eb536fe21eb, type: 3}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
